backend/requirements.txt
# Django
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
django-celery-beat==2.7.0
django-celery-results==2.5.1
daphne==4.1.2
channels==4.2.0
channels-redis==4.2.1

# Database
psycopg2-binary==2.9.10
django-environ==0.11.2

# Async tasks
celery==5.4.0
redis==5.2.1
flower==2.0.1

# AI/ML
openai==1.55.3
langchain==0.3.9
langchain-openai==0.2.11
langchain-community==0.3.9
langchain-experimental==0.3.3
langchain-chroma==0.1.4
chromadb==0.5.23

# Data Science
pandas==2.2.3
numpy==1.26.4
scipy==1.14.1
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1
statsmodels==0.14.4

# ML Frameworks
xgboost==2.1.3
torch==2.5.1 --index-url https://download.pytorch.org/whl/cpu
tensorflow-cpu==2.18.0

# Document Processing
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2

# Utilities
python-dotenv==1.0.1
httpx==0.28.0
aiohttp==3.11.9
pydantic==2.10.2
tiktoken==0.8.0
python-magic==0.4.27
Pillow==11.0.0
kaggle==1.6.17
google-search-results==2.4.2

# Code execution sandbox
RestrictedPython==7.4

# API documentation
drf-spectacular==0.28.0

# Testing
pytest==8.3.3
pytest-django==4.9.0
pytest-asyncio==0.24.0
factory-boy==3.3.1
